Move formatter to core module
