+++ core/src/main/kotlin/ru/mobileup/template/core/utils/HtmlUtils.kt
@@ -1,10 +1,10 @@
-package ru.mobileup.template.features.coins.data.dto
+package ru.mobileup.template.core.utils
 
 import android.os.Build
 import android.text.Html
 
 @Suppress("DEPRECATION")
-internal fun String.formatDescription() = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+fun String.fromHtml() = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
     Html.fromHtml(this, Html.FROM_HTML_MODE_LEGACY)
 } else {
     Html.fromHtml(this)
